refactor buy/sell modules into separate directories
