+++ server/src/trade/buy/buy.config.json
@@ -1,0 +1,3 @@
+{
+    "strategies": []
+}
