+++ test/unit/server/trade/buy.test.ts
@@ -1,4 +1,4 @@
-const { buySymbols, getBuyableSymbols, buy } = require('../../../../server/src/trade/buy');
+const { buySymbols, getBuyableSymbols, buy } = require('../../../../server/src/trade/buy/buy.ts');
 
 describe('Buy module', () => {
     describe('buySymbols()', () => {
