+++ test/unit/server/trade/sell.test.ts
@@ -1,4 +1,4 @@
-const { sellPositions, getPositions, isPositionSellable, sellPosition } = require('../../../../server/src/trade/sell');
+const { sellPositions, getPositions, isPositionSellable, sellPosition } = require('../../../../server/src/trade/sell/sell.ts');
 
 describe('Sell module', () => {
     describe('sellPositions()', () => {
